feat: USD/Alembic export, shared intersection predicate, and CPU eigensolver
- Add USD/Alembic cache export operators to the Blender addon as a lighter
  alternative to the shape-key bake, with a debug scenario for coverage
- Share the edge-triangle pierce predicate between the GPU solver and the
  Rust build-time self-intersection check via a C ABI (intersect_ffi.cpp,
  intersect_core.hpp, build.rs) so both run the same geometry
- Add a CPU linalg eigensolver (linalg/eigsolve.hpp, smat.hpp) and extend
  eigenanalysis, solver, and energy models

+++ crates/ppf-cts-solver/src/cpp/barrier/logarithm.hpp
@@ -17,7 +17,7 @@ __device__ static float energy(float g, float ghat, float offset) {
     } else if (g >= ghat) {
         return 0.0f;
     }
-    return -(g - ghat) * (g - ghat) * log(g / ghat);
+    return -(g - ghat) * (g - ghat) * logf(g / ghat);
 }
 
 __device__ static float gradient(float g, float ghat, float offset) {
@@ -27,7 +27,7 @@ __device__ static float gradient(float g, float ghat, float offset) {
     } else if (g >= ghat) {
         return 0.0f;
     }
-    return (ghat - g) * (2.0f * g * log(g / ghat) + g - ghat) / g;
+    return (ghat - g) * (2.0f * g * logf(g / ghat) + g - ghat) / g;
 }
 
 __device__ static float curvature(float g, float ghat, float offset) {
@@ -37,7 +37,7 @@ __device__ static float curvature(float g, float ghat, float offset) {
     } else if (g >= ghat) {
         return 0.0f;
     }
-    return -2.0f * log(g / ghat) + ghat * (ghat + 2.0f * g) / (g * g) - 3.0f;
+    return -2.0f * logf(g / ghat) + ghat * (ghat + 2.0f * g) / (g * g) - 3.0f;
 }
 
 } // namespace logarithm
